Modularize mock of Performance host object
